+++ Code/Data Classes/SpecialAbilityData.cs
@@ -1,0 +1,10 @@
+﻿
+namespace JLPlugin.Data
+{
+    [System.Serializable]
+    public class SpecialAbilityData
+    {
+        public string name;
+        public string GUID;
+    }
+}